# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PPZ SPA eInvoice API Documentation",
        "description": "API documentation for PPZ SPA eInvoice system",
        "version": "1.0.0"
    },
    "paths": {
        "/api/spa/einvoice/login-taxpayer": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Login taxpayer and get access token",
                "operationId": "1d2e5b7f9484dd21c22569175ffdab6e",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "type": "string",
                                            "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJSUzI1NiIs..."
                                        },
                                        "expires_in": {
                                            "type": "integer",
                                            "example": 3600
                                        },
                                        "token_type": {
                                            "type": "string",
                                            "example": "Bearer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "Server Error"
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Authentication",
            "description": "Authentication"
        }
    ]
}